requirements.txt
discord.py>=2.0.0
aiohttp<4,>=3.7.4
attrs>=17.3.0
aiosignal>=1.1.2
async-timeout<5.0,>=4.0.0a3
yarl<2.0,>=1.0
charset-normalizer<3.0,>=2.0
multidict<7.0,>=4.5
frozenlist>=1.1.1
idna>=2.0